+++ src/views/default/create.php
@@ -9,16 +9,14 @@ use yii\bootstrap\Alert;
 use yii\helpers\Html;
 use yii\widgets\ActiveForm;
 use ymaker\email\templates\Module as TemplatesModule;
-use ymaker\email\templates\widgets\LanguagesList;
+use ymaker\email\templates\helpers\LanguageHelper;
 use vova07\imperavi\Widget as ImperaviRedactor;
 
 /**
  * View file for CRUD backend controller.
  *
  * @var \yii\web\View $this
- * @var \ymaker\email\templates\models\entities\EmailTemplate $template
- * @var \ymaker\email\templates\models\entities\EmailTemplateTranslation $translation
- * @var array $errors
+ * @var \ymaker\email\templates\models\entities\EmailTemplate $model
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
  * @since 1.0
@@ -38,38 +36,24 @@ $this->params['breadcrumbs'][] = TemplatesModule::t('Create email template');
             <h1>
                 <?= TemplatesModule::t('Email templates') ?>
                 <small><?= TemplatesModule::t('create new template') ?></small>
-                <div class="pull-right">
-                    <?= LanguagesList::widget(['currentLanguage' => $translation->language]) ?>
-                </div>
             </h1>
         </div>
         <div class="clearfix"></div>
         <hr>
-        <div class="col-md-12">
-            <?php if (isset($errors)): ?>
-                <?php foreach ($errors as $fieldErrors): ?>
-                    <?php foreach ($fieldErrors as $error): ?>
-                        <?= Alert::widget([
-                            'body' => $error,
-                            'options' => [
-                                'class' => 'alert-danger'
-                            ],
-                        ]) ?>
-                    <?php endforeach; ?>
-                <?php endforeach; ?>
-            <?php endif; ?>
-        </div>
         <div class="col-md-12">
             <?php $form = ActiveForm::begin([
                 'fieldConfig' => [
                     'template' => "{label}\n{input}\n{error}",
                 ],
             ]) ?>
-            <?= $form->field($template, 'key')->textInput(['autofocus' => true]) ?>
-            <?= $form->field($translation, 'subject') ?>
-            <?= $form->field($translation, 'body')->widget(ImperaviRedactor::class) ?>
-            <?= $form->field($translation, 'hint') ?>
-            <?= $form->field($translation, 'language')->hiddenInput()->label(false) ?>
+            <?= $form->field($model, 'key')->textInput(['autofocus' => true]) ?>
+            <?php foreach (LanguageHelper::getLocales() as $language): ?>
+                <?php $translation = $model->getTranslation($language) ?>
+                <?= $form->field($translation, 'subject') ?>
+                <?= $form->field($translation, 'body')
+                    ->widget(ImperaviRedactor::class) ?>
+                <?= $form->field($translation, 'hint') ?>
+            <?php endforeach ?>
             <?= Html::submitButton(
                 TemplatesModule::t('Create'),
                 ['class' => 'btn btn-success']