+++ src/views/default/update.php
@@ -9,17 +9,14 @@ use yii\bootstrap\Alert;
 use yii\helpers\Html;
 use yii\widgets\ActiveForm;
 use ymaker\email\templates\Module as TemplatesModule;
-use ymaker\email\templates\widgets\LanguagesList;
+use ymaker\email\templates\helpers\LanguageHelper;
 use vova07\imperavi\Widget as ImperaviRedactor;
 
 /**
  * View file for CRUD backend controller.
  *
  * @var \yii\web\View $this
- * @var \ymaker\email\templates\models\entities\EmailTemplate $template
- * @var \ymaker\email\templates\models\entities\EmailTemplateTranslation $translation
- * @var \ymaker\email\templates\models\entities\EmailTemplateTranslation $defaultTranslation
- * @var array $errors
+ * @var \ymaker\email\templates\models\entities\EmailTemplate $model
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
  * @since 1.0
@@ -30,7 +27,7 @@ $this->params['breadcrumbs'][] = [
     'url' => ['/email-templates/default/index'],
 ];
 $this->params['breadcrumbs'][] = TemplatesModule::t('Update email template - {key}', [
-    'key' => $template->key,
+    'key' => $model->key,
 ]);
 
 \yii\bootstrap\BootstrapAsset::register($this);
@@ -41,45 +38,23 @@ $this->params['breadcrumbs'][] = TemplatesModule::t('Update email template - {ke
             <h1>
                 <?= TemplatesModule::t('Email templates') ?>
                 <small><?= TemplatesModule::t('update template') ?></small>
-                <div class="pull-right">
-                    <?= LanguagesList::widget(['currentLanguage' => $translation->language]) ?>
-                </div>
             </h1>
         </div>
         <div class="clearfix"></div>
         <hr>
         <div class="col-md-12">
-            <?php if (isset($errors)): ?>
-                <?php foreach ($errors as $fieldErrors): ?>
-                    <?php foreach ($fieldErrors as $error): ?>
-                        <?= Alert::widget([
-                            'body' => $error,
-                            'options' => [
-                                'class' => 'alert-danger'
-                            ],
-                        ]) ?>
-                    <?php endforeach; ?>
-                <?php endforeach; ?>
-            <?php endif; ?>
-        </div>
-        <div class="col-md-12">
             <?php $form = ActiveForm::begin() ?>
-            <?= $form->field($template, 'key')
+            <?= $form->field($model, 'key')
                 ->textInput(['disabled' => true]) ?>
-            <?= $form->field($translation, 'subject')
-                ->textInput(['autofocus' => true])
-                ->hint($defaultTranslation->getAttributeHint('subject')) ?>
-            <?= $form->field($translation, 'body')
-                ->widget(ImperaviRedactor::class)
-                ->hint($defaultTranslation->getAttributeHint('body')) ?>
-            <?= $form->field($translation, 'hint')
-                ->textInput(['disabled' => !$translation->getIsNewRecord()]) ?>
-            <?= $form->field($translation, 'templateId')
-                ->hiddenInput()
-                ->label(false) ?>
-            <?= $form->field($translation, 'language')
-                ->hiddenInput()
-                ->label(false) ?>
+            <?php foreach (LanguageHelper::getLocales() as $language): ?>
+                <?php $translation = $model->getTranslation($language) ?>
+                <?= $form->field($translation, 'subject')
+                    ->textInput() ?>
+                <?= $form->field($translation, 'body')
+                    ->widget(ImperaviRedactor::class) ?>
+                <?= $form->field($translation, 'hint')
+                    ->textInput(['disabled' => true]) ?>
+            <?php endforeach ?>
             <?= Html::submitButton(
                 TemplatesModule::t('Save'),
                 ['class' => 'btn btn-success']